backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
celery==5.3.4
redis==5.0.1
langchain==0.0.348
openai==1.3.7
transformers==4.36.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1